encapsulating sendmessage
